refactor: migrate hooks (use-debounced-duplicate-issues) from web/app/ce to web/app/core
